Unify the GithubReleaseVersionManager between normal and enterprise versions
