--- kmmbridge/src/main/kotlin/artifactmanager/GithubReleaseException.kt
@@ -1,3 +1,0 @@
-package artifactmanager
-
-class GithubReleaseException(message: String, cause: Throwable? = null) : Exception(message, cause)
