+++ kmmbridge/src/main/kotlin/internal/GithubCalls.kt
@@ -13,8 +13,6 @@
 
 package co.touchlab.faktory.internal
 
-import artifactmanager.GithubApi
-import artifactmanager.GithubReleaseException
 import co.touchlab.faktory.githubPublishToken
 import com.google.gson.Gson
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
